backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-github-s3==0.25
djangorestframework==3.14.0
Faker==19.3.1
faker-vehicle==0.2.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.4
stripe==6.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0